+++ packages/pluggableWidgets/html-element-web/CHANGELOG.md
@@ -8,4 +8,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   We introduced HTML Node widget!
+-   We introduced HTML Element widget!
