+++ packages/pluggableWidgets/html-element-web/README.md
@@ -1,0 +1,3 @@
+# HTML Element
+
+Please see [HTML Element](https://docs.mendix.com/appstore/widgets/html-element) in the Mendix documentation for details.
