+++ packages/pluggableWidgets/html-element-web/src/HTMLElement.editorPreview.tsx
@@ -1,0 +1,11 @@
+import { ReactElement, createElement } from "react";
+import { HTMLElementPreviewProps } from "../typings/HTMLElementProps";
+
+export function preview(props: HTMLElementPreviewProps): ReactElement {
+    return <div className={props.className}>HTML Element</div>;
+}
+
+export function getPreviewCss(): string {
+    // html element has no styling by design
+    return "";
+}
