+++ packages/pluggableWidgets/html-element-web/src/utils/style-utils.ts
