--- packages/pluggableWidgets/html-node-web/README.md
@@ -1,3 +1,0 @@
-# HTMLNode
-
-Please see [HTMLNode](https://docs.mendix.com/appstore/widgets/htmlnode) in the Mendix documentation for details.
