--- packages/pluggableWidgets/html-node-web/src/package.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<package xmlns="http://www.mendix.com/package/1.0/">
-    <clientModule name="HTMLNode" version="1.0.0" xmlns="http://www.mendix.com/clientModule/1.0/">
-        <widgetFiles>
-            <widgetFile path="HTMLNode.xml" />
-        </widgetFiles>
-        <files>
-            <file path="com/mendix/widget/web/htmlnode/" />
-        </files>
-    </clientModule>
-</package>
